[DEV-2028] Refactoring
