+++ src/subdomains/generic/kyc/controllers/kyc.controller.ts
@@ -30,9 +30,8 @@ import { CountryDto } from 'src/shared/models/country/dto/country.dto';
 import { DfxLogger } from 'src/shared/services/dfx-logger';
 import { IdentStatus } from '../dto/ident.dto';
 import { IdentResultDto } from '../dto/input/ident-result.dto';
-import { KycContactData } from '../dto/input/kyc-contact-data.dto';
+import { KycContactData, KycPersonalData } from '../dto/input/kyc-data.dto';
 import { KycFinancialInData } from '../dto/input/kyc-financial-in.dto';
-import { KycPersonalData } from '../dto/input/kyc-personal-data.dto';
 import { Verify2faDto } from '../dto/input/verify-2fa.dto';
 import { KycFinancialOutData } from '../dto/output/kyc-financial-out.dto';
 import { KycSessionDto, KycStatusDto } from '../dto/output/kyc-info.dto';
