--- src/subdomains/generic/kyc/dto/input/kyc-contact-data.dto.ts
@@ -1,9 +1,0 @@
-import { ApiProperty } from '@nestjs/swagger';
-import { IsEmail, IsNotEmpty } from 'class-validator';
-
-export class KycContactData {
-  @ApiProperty()
-  @IsNotEmpty()
-  @IsEmail()
-  mail: string;
-}
